# docs/swagger/swagger.yaml
basePath: /api/v1/
definitions:
  domain.Location:
    properties:
      latitude:
        type: number
      longitude:
        type: number
    type: object
  domain.Restaurant:
    properties:
      id:
        type: integer
      image:
        type: string
      location:
        $ref: '#/definitions/domain.Location'
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      working_status:
        type: string
    type: object
  v1.response:
    properties:
      message:
        type: string
    type: object
  v1.signInInput:
    properties:
      name:
        type: string
      password:
        type: string
    type: object
  v1.tokenResponse:
    properties:
      token:
        type: string
    type: object
  v1.userSignUpInput:
    properties:
      email:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      password:
        type: string
      phone:
        type: string
    type: object
host: localhost:9000
info:
  contact: {}
  description: API Server for Yet Another Delivery App
  title: Yet Another Delivery API
  version: "1.0"
paths:
  /admins/sign-in:
    post:
      consumes:
      - application/json
      description: admin sign in
      parameters:
      - description: sign up info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.signInInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.tokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      summary: Admin SignIn
      tags:
      - admins-auth
  /couriers/sign-in:
    post:
      consumes:
      - application/json
      description: courier sign in
      parameters:
      - description: sign up info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.signInInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.tokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      summary: Courier SignIn
      tags:
      - couriers
  /restaurants:
    get:
      consumes:
      - application/json
      description: get all restaurant for user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Restaurant'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "403":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - UserAuth: []
      summary: Get All Restaurant
      tags:
      - restaurants
  /restaurants/{rid}:
    get:
      consumes:
      - application/json
      description: get restaurant by id for user
      parameters:
      - description: Restaurant id
        in: path
        name: rid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Restaurant'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "403":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - UserAuth: []
      summary: Get Restaurant By Id
      tags:
      - restaurants
  /restaurants/sign-in:
    post:
      consumes:
      - application/json
      description: restaurant sign in
      parameters:
      - description: sign up info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.signInInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.tokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      summary: Restaurants SignIn
      tags:
      - restaurants
  /users/sign-in:
    post:
      consumes:
      - application/json
      description: user sign in
      parameters:
      - description: sign up info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.signInInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.tokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      summary: User SignIn
      tags:
      - users-auth
  /users/sign-up:
    post:
      consumes:
      - application/json
      description: user sign up
      parameters:
      - description: sign up info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.userSignUpInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.tokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      summary: User SignUp
      tags:
      - users-auth
securityDefinitions:
  AdminAuth:
    in: header
    name: Authorization
    type: apiKey
  CourierAuth:
    in: header
    name: Authorization
    type: apiKey
  RestaurantAuth:
    in: header
    name: Authorization
    type: apiKey
  UserAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"